[MID-39] add Event.isActive, imports refactoring
