+++ jsconfig.json
@@ -6,6 +6,7 @@
       "@images/*": ["./public/assets/images/*"],
       "@public/*": ["./public/assets/*"],
       "@components": ["./src/app/_components/index"],
+      "@components/*": ["./src/app/_components/*"],
       "@hooks": ["./src/app/_hooks/index"],
       "@admin/*": ["./src/app/(admin)/admin/*"],
       "@admin/components/*": ["./src/app/(admin)/admin/_components/*"],
