+++ src/app/(app)/page.js
@@ -4,7 +4,6 @@ import { prisma } from '@/lib/db';
 // Page metadata should contain
 // title - gets formatted into "%s | Маяк", %s is replaced by title,
 // description - short description of the page,
-
 export const metadata = {
   title: 'Головна сторінка',
   description: 'Пошук психологічної допомоги в м.Львів',
