+++ src/app/_components/CheckBox/components.js
@@ -1,6 +1,5 @@
 'use client';
 
-import React from 'react';
 import PropTypes from 'prop-types';
 import TickIcon from '@icons/tickIcon.svg';
 import { cn } from '@/utils/cn';
