+++ src/app/_components/CheckBox/index.js
@@ -1,1 +1,1 @@
-export { CheckBox } from './CheckBox';
+export * from './components';
