--- src/app/_components/Feedback/index.js
@@ -1,1 +1,0 @@
-export * from './Feedback';
