+++ src/app/_components/Links/SocialLink.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import p from 'prop-types';
 import Link from 'next/link';
 import { cn } from '@utils/cn';
