--- src/app/_components/Therapies/index.js
@@ -1,1 +1,0 @@
-export * from './Therapies';
