+++ src/app/_components/index.js
@@ -1,14 +1,13 @@
-export * from './InputFields';
-export * from './Links';
-export * from './Modal';
-export * from './PillButton';
-export * from './Typography';
-export * from './Footer';
-export * from './Header';
-export * from './OutlinedButton';
-export * from './TextArea';
-export * from './TopWave';
-export * from './BottomWave';
-export * from './TherapyCard';
-export * from './TherapiesSection';
-export * from './Therapies';
+export * from '@components/InputFields';
+export * from '@components/Links';
+export * from '@components/Modal';
+export * from '@components/PillButton';
+export * from '@components/Typography';
+export * from '@components/Footer';
+export * from '@components/Header';
+export * from '@components/OutlinedButton';
+export * from '@components/TextArea';
+export * from '@components/TopWave';
+export * from '@components/BottomWave';
+export * from '@components/TherapyCard';
+export * from '@components/TherapiesSection';
